+++ CHANGELOG.md
@@ -7,6 +7,8 @@ clj has both stable and prerelease versions. Current and former stable build are
 
 Prerelease versions:
 
+* next
+  * On -X success, exit with (shutdown-agents) instead of System/exit
 * 1.10.3.905 on July 9, 2021
   * Removing multi-function support from -X and -T for now
 * 1.10.3.899 on July 9, 2021